feat: first version with task analysis
- Создана команда для получения отделов из битрикс24
- Создана команда для получения всех пользователей из битрикс24
- Создана команда для получения задач пользователей из битрикс24
- Добавлен триггер для отправки задач ChatGPT и получения от него анализа

+++ database/migrations/2024_12_25_090120_create_tasks_table.php
@@ -23,7 +23,7 @@ return new class extends Migration
             $table->unsignedBigInteger('closed_by')->nullable();
             $table->timestamp('deadline')->nullable();
             $table->unsignedTinyInteger('status');
-            $table->unsignedTinyInteger('sub_status');
+            $table->smallInteger('sub_status');
             $table->timestamps();
 
             $table->foreign('created_by')->references('id')->on('bx_users');
